--- subprojects/platform-jvm/src/main/java/org/gradle/jvm/internal/apigen/abi/AnnotationAnnotationValue.java
@@ -1,29 +1,0 @@
-/*
- * Copyright 2015 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.gradle.jvm.internal.apigen.abi;
-
-public class AnnotationAnnotationValue extends AnnotationValue {
-    private final AnnotationSig annotation;
-
-    public AnnotationAnnotationValue(String name, AnnotationSig annotation) {
-        super(name);
-        this.annotation = annotation;
-    }
-
-    public AnnotationSig getAnnotation() {
-        return annotation;
-    }
-}
