+++ subprojects/platform-jvm/src/main/java/org/gradle/jvm/tasks/api/AlwaysMemberOfApiChecker.java
@@ -13,17 +13,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.gradle.jvm.internal.apigen.abi;
 
-public class EnumAnnotationValue extends SimpleAnnotationValue {
-    private final String desc;
+package org.gradle.jvm.tasks.api;
 
-    public EnumAnnotationValue(String name, String desc, String value) {
-        super(name, value);
-        this.desc = desc;
-    }
-
-    public String getDesc() {
-        return desc;
+public class AlwaysMemberOfApiChecker implements MemberOfApiChecker {
+    @Override
+    public boolean belongsToApi(String className) {
+        return true;
     }
 }
