+++ subprojects/platform-jvm/src/main/java/org/gradle/jvm/tasks/api/AnnotationSig.java
@@ -1,0 +1,54 @@
+/*
+ * Copyright 2015 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.gradle.jvm.tasks.api;
+
+import com.google.common.collect.ComparisonChain;
+import com.google.common.collect.Lists;
+
+import java.util.List;
+
+public class AnnotationSig implements Comparable<AnnotationSig> {
+    private final String name;
+    private final boolean visible;
+    private final List<AnnotationValue> values = Lists.newLinkedList();
+
+    public AnnotationSig(String name, boolean visible) {
+        this.name = name;
+        this.visible = visible;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public List<AnnotationValue> getValues() {
+        return values;
+    }
+
+    public boolean isVisible() {
+        return visible;
+    }
+
+    @Override
+    public int compareTo(AnnotationSig o) {
+        return ComparisonChain.start()
+            .compare(name, o.name)
+            .compare(visible, o.visible)
+            .result();
+    }
+
+}
