+++ subprojects/platform-jvm/src/main/java/org/gradle/jvm/tasks/api/MemberOfApiChecker.java
@@ -13,17 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.gradle.jvm.internal.apigen.abi;
 
-public class ArrayAnnotationValue extends AnnotationValue {
-    private final AnnotationValue[] value;
+package org.gradle.jvm.tasks.api;
 
-    public ArrayAnnotationValue(String name, AnnotationValue[] value) {
-        super(name);
-        this.value = value;
-    }
-
-    public AnnotationValue[] getValue() {
-        return value;
-    }
+public interface MemberOfApiChecker {
+    boolean belongsToApi(String className);
 }
