+++ subprojects/platform-jvm/src/main/java/org/gradle/jvm/tasks/api/SimpleAnnotationValue.java
@@ -13,10 +13,18 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.gradle.jvm.internal.apigen.abi;
 
-public class InvalidPublicAPIException extends RuntimeException {
-    public InvalidPublicAPIException(String message) {
-        super(message);
+package org.gradle.jvm.tasks.api;
+
+public class SimpleAnnotationValue extends AnnotationValue {
+    private final Object value;
+
+    public SimpleAnnotationValue(String name, Object value) {
+        super(name);
+        this.value = value;
+    }
+
+    public Object getValue() {
+        return value;
     }
 }
